Fix ZIP file timestamps and compilation warnings
Fixed issues:
1. Timestamps now correctly set from file modification time
2. Fixed type mismatch (uint32_t -> uint16_t) for DOS date/time fields
3. Eliminated packed struct pointer warnings
4. Added tmp/ directory for test artifacts

Changes:
- Store timestamps in file_entry and pass to write_local_header()
- Use uint16_t for last_mod_time/date in file_entry struct
- Add tmp/ to .gitignore for test files

Testing:
- Dates now show correctly: 2025-12-22 19:15
- No compilation warnings
- Archives still validate with unzip

+++ include/burst_writer.h
@@ -24,8 +24,8 @@ struct file_entry {
     uint16_t compression_method;
     uint16_t version_needed;
     uint16_t general_purpose_flags;
-    uint32_t last_mod_time;
-    uint32_t last_mod_date;
+    uint16_t last_mod_time;
+    uint16_t last_mod_date;
 };
 
 // BURST writer context
